requirements.txt
numpy==1.26.1
pandas==2.1.1
matplotlib==3.8.0
seaborn==0.13.2
plotly==5.17.0
Pillow==10.0.1
streamlit==1.40.2
joblib==1.4.2
scikit-learn==1.3.1
tensorflow-cpu==2.16.1
keras>=3.0.0
gdown